Practicum1
I created an outdoor environment of a garden and a road. There are a table, several lights, a tree, a box and a sphere in the garden. The box can be picked up, and the sphere is changing color. The color of the skybox is changed to a warmer color and the sun is rotating to show the time changing.  There are five sounds added to the environment. The sound of rain, wind, tree and two lights.